backend/requirements.txt
# ============================================================================
# FAANG-Grade Dependencies for Landing Zone Portal Backend
# ============================================================================

# Core Framework
FastAPI==0.109.0
pydantic==2.5.0
pydantic-settings==2.1.0
starlette==0.35.1

# ASGI Server
uvicorn[standard]==0.27.0
gunicorn==21.2.0
uvloop==0.19.0
httptools==0.6.1

# Google Cloud SDK
google-cloud-firestore==2.14.0
google-cloud-bigquery==3.14.0
google-cloud-logging==3.9.0
google-cloud-pubsub==2.19.0
google-cloud-secret-manager==2.17.0
google-cloud-resource-manager>=1.11.0
google-cloud-asset>=3.22.0
google-cloud-monitoring>=2.18.0
google-cloud-storage>=2.14.0

# Authentication
google-auth>=2.25.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
python-json-logger==2.0.7

# Caching (Redis)
redis>=5.0.0
aioredis>=2.0.1

# Observability
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-exporter-gcp-trace>=1.6.0
opentelemetry-propagator-gcp>=1.6.0
prometheus-client>=0.19.0

# Security
cryptography>=41.0.7
PyJWT>=2.8.0

# ============================================================================
# Testing
# ============================================================================
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
pytest-mock==3.12.0
httpx==0.26.0
respx>=0.20.2
factory-boy>=3.3.0
faker>=22.0.0

# ============================================================================
# Code Quality
# ============================================================================
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
bandit==1.7.6
# safety removed due to dependency conflict with packaging

# Type stubs
types-redis>=4.6.0
types-python-dateutil>=2.8.19

# ============================================================================
# Development
# ============================================================================
watchfiles==0.21.0
ipython>=8.18.0
rich>=13.7.0
